Move event to its own module
