+++ src/index.rs
@@ -30,9 +30,13 @@ use {
   },
 };
 
-pub use {self::entry::RuneEntry, entry::MintEntry};
+pub use {
+  self::{entry::RuneEntry, event::Event},
+  entry::MintEntry,
+};
 
 pub(crate) mod entry;
+mod event;
 mod fetcher;
 mod reorg;
 mod rtx;
@@ -195,29 +199,6 @@ impl<T> BitcoinCoreRpcResultExt<T> for Result<T, bitcoincore_rpc::Error> {
       Err(err) => Err(err.into()),
     }
   }
-}
-
-/// An event from indexing which can be optionally emitted by setting a
-/// channel sender using `set_event_sender`.
-#[derive(Debug, Clone)]
-pub enum Event {
-  /// Newly created inscriptions will include additional metadata including
-  /// rarity, cursed status, charms, etc.
-  InscriptionCreated {
-    id: InscriptionId,
-    location: Option<SatPoint>,
-    sequence_number: u32,
-    block_height: u32,
-    charms: u16,
-    parent_inscription_id: Option<InscriptionId>,
-  },
-  InscriptionMoved {
-    id: InscriptionId,
-    old_location: SatPoint,
-    new_location: SatPoint,
-    sequence_number: u32,
-    block_height: u32,
-  },
 }
 
 pub struct Index {
